Move Logger to its own file.
