+++ mathematica/mathematica.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <filesystem>
 #include <map>
 #include <string>
 #include <tuple>
@@ -8,7 +7,6 @@
 #include <vector>
 
 #include "astronomy/orbital_elements.hpp"
-#include "base/file.hpp"
 #include "base/traits.hpp"
 #include "geometry/grassmann.hpp"
 #include "geometry/point.hpp"
@@ -32,7 +30,6 @@ namespace mathematica {
 namespace internal_mathematica {
 
 using base::all_different_v;
-using base::OFStream;
 using geometry::Bivector;
 using geometry::Point;
 using geometry::Quaternion;
@@ -62,10 +59,6 @@ using quantities::Quantity;
 using quantities::Quotient;
 using quantities::Temperature;
 using quantities::Time;
-
-// Define this value to 1 to force the logger to append "_new" to the file
-// names, which is useful for regression testing of the logger.
-#define PRINCIPIA_MATHEMATICA_LOGGER_REGRESSION_TEST 0
 
 // A helper class for type erasure of quantities.  It may be used with the
 // functions in this file to remove the dimensions of quantities (we know that
@@ -284,50 +277,12 @@ template<typename OptionalExpressIn = std::nullopt_t>
 std::string ToMathematica(std::string const& str,
                           OptionalExpressIn express_in = std::nullopt);
 
-// An RAII object to help with Mathematica logging.
-class Logger final {
- public:
-  // Creates a logger object that will, at destruction, write to the given file.
-  // If make_unique is true, a unique id is inserted before the file extension
-  // to identify different loggers.
-  Logger(std::filesystem::path const& path,
-         bool make_unique = true);
-  ~Logger();
-
-  // Flushes the contents of the logger to the file.  This should not normally
-  // be called explicitly, but it's useful when debugging crashes.  Beware, the
-  // resulting file may contain many assignments to the same variable.
-  void Flush();
-
-  // Appends an element to the list of values for the List variable |name|.  The
-  // |args...| are passed verbatim to ToMathematica for stringification.  When
-  // this object is destroyed, an assignment is generated for each of the
-  // variables named in a call to Append.
-  template<typename... Args>
-  void Append(std::string const& name, Args... args);
-
-  // Sets an element as the single value for the variable |name|.  The
-  // |args...| are passed verbatim to ToMathematica for stringification.  When
-  // this object is destroyed, an assignment is generated for each of the
-  // variables named in a call to Set.
-  template<typename... Args>
-  void Set(std::string const& name, Args... args);
-
- private:
-  OFStream file_;
-  std::map<std::string, std::vector<std::string>> name_and_multiple_values_;
-  std::map<std::string, std::string> name_and_single_value_;
-
-  static std::atomic_uint64_t id_;
-};
-
 }  // namespace internal_mathematica
 
 using internal_mathematica::Apply;
 using internal_mathematica::Assign;
 using internal_mathematica::Evaluate;
 using internal_mathematica::ExpressIn;
-using internal_mathematica::Logger;
 using internal_mathematica::Option;
 using internal_mathematica::PlottableDataset;
 using internal_mathematica::ToMathematica;
